IT'S ALIVE!
Version 0.1 of the scheduling tool is complete! It outputs two files: event_calendar.csv, which is the list of all the events in Google Calendar form, and assigned_mentors.csv, which is the list of all the events and which mentor was assigned to it. This assigned_mentors file should be made clearer in the future.